requirements.txt
pytest==6.2.5
requests==2.32.0
python-dotenv==0.20.0